refactor to only generate 1.1
